Preserve Nim int type for minimum literals

+++ tests/integration/cases/scalar_int_i64_min/Nim_i64_min_integer_format_hex@v2.nim
@@ -1,2 +1,2 @@
 import json
-var my_data = %* cast[int64](0x8000000000000000'u64)
+var my_data = %* cast[int](0x8000000000000000'u64)
