+++ tests/integration/cases/scalar_int_i64_min/Nim_i64_min_integer_format_octal@v2.nim
@@ -1,2 +1,2 @@
 import json
-var my_data = %* cast[int64](0o1000000000000000000000'u64)
+var my_data = %* cast[int](0o1000000000000000000000'u64)
